% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{make_long}
\alias{make_long}
\title{make_long}
\usage{
make_long(.df, ...)
}
\arguments{
\item{.df}{a data frame}

\item{...}{unquoted columnnames of df that you want to include in the plot.}
}
\value{
a longer data frame
}
\description{
Prepares a 'wide' data frame into a format that `geom_sankey` or `geom_alluvial` understands. Useful to show flows between dimensions in dataset.
}
